client/requirements.txt
appdirs==1.4.3
backoff==1.10.0
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
googleapis-common-protos==1.52.0
grpcio==1.34.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
opentelemetry-api==0.16b1
opentelemetry-exporter-otlp==0.16b1
opentelemetry-instrumentation==0.16b1
opentelemetry-instrumentation-dbapi==0.16b1
opentelemetry-instrumentation-grpc==0.16b1
opentelemetry-instrumentation-requests==0.16b1
opentelemetry-instrumentation-sqlite3==0.16b1
opentelemetry-proto==0.16b1
opentelemetry-sdk==0.16b1
packaging==20.3
pep517==0.8.2
progress==1.5
protobuf==3.14.0
pyparsing==2.4.6
pytoml==0.1.21
requests==2.22.0
retrying==1.3.3
six==1.14.0
urllib3==1.25.8
webencodings==0.5.1
wrapt==1.12.1
